requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
python-dotenv==1.0.0
jose==1.6.0
httpx==0.24.0